+++ app/main/views/create_a_requirement/__init__.py
@@ -1,1 +1,2 @@
 from . import create
+from . import publish
